Fix pip install
`pip install -e .` and `python setup.py develop` worked properly,
but `pip install .` and `pip install [link]` did not.
The find_packages() command in setup.py looks for `__init__.py`.
Due to the recent reorganization of the library, there was no
`__init__.py` in the `gadal` folder. This is not compliant with
pre-3.3 packages and is generally not preferred.

Furthermore, the unit tests required that the `tests` package be installed.
This is not removed and the `__init__.py` code in `tests` has now been
moved into the individual unit testing files.